# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wallet Tracker REST API",
    "version": "1.0.0",
    "description": "Contract for the explicit REST operations implemented by the current server: health, authentication, categories, transactions, synthetic batch import, schedules, summary, settings and AI drafts. Descriptions record implementation rules that JSON Schema cannot fully express. Document validation alone is not evidence of executed HTTP scenarios: check:openapi validates this contract and test:contract separately exercises disposable offline/demo servers, reporting observed and unexercised statuses. No full-project regression or GitHub CI result is implied by this document. OpenAPI 3.0.3 is selected for parser compatibility, not claimed to be the latest version. This describes a personal project, not a production service."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local application; PORT is configurable. No hosted demo is promised."
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/transactions": {
      "get": {
        "operationId": "listTransactions",
        "tags": ["Transactions"],
        "summary": "List own expenses within an inclusive local-date range",
        "description": "Defaults from/to independently to current server-local month's boundaries; real calendar dates required and from<=to. Sorted spent_on descending, then id descending. limit defaults50, accepts0-200 after Number conversion; offset defaults0, accepts nonnegative integers with no explicit ceiling. Empty limit/offset use defaults. Optional category_id uses ownership validation. Empty category_id maps null, but SQL category_id=null matches no rows; it is not an uncategorised filter. Unknown query keys ignored.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "name": "from", "in": "query", "schema": { "$ref": "#/components/schemas/LocalDate" }, "description": "Inclusive start, omitted=current month start." },
          { "name": "to", "in": "query", "schema": { "$ref": "#/components/schemas/LocalDate" }, "description": "Inclusive end, omitted=current month end." },
          { "name": "category_id", "in": "query", "schema": { "type": "string" }, "description": "Number-converted owned positive integer ID; foreign/missing404, invalid400. Empty string mapsnull and matches no rows." },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "50" }, "description": "Number(value) must be integer0-200; empty string defaults50." },
          { "name": "offset", "in": "query", "schema": { "type": "string", "default": "0" }, "description": "Number(value) must be nonnegative integer; empty string defaults0. No explicit upper ceiling." }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/TransactionList" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createTransaction",
        "tags": ["Transactions"],
        "summary": "Save an expense, optionally with an account-scoped operation key",
        "description": "Positive integer cents only. category_id uses Number conversion and own-category lookup; note uses String(non-null JSON). Omitted/null unkeyed spent_on defaults to server-local today. New expenses must use a real date no later than tomorrow; no explicit lower-year bound. Extra POST fields ignored. Budgets never block saving. With Idempotency-Key, spent_on must be explicit/non-null; canonical [amount,category-or-null,date,note-or-null] identifies the payload, independent of JSON key order and ignored extras. New keyed saves recheck ownership/date bound atomically. Same-key/same-payload replay returns original201 snapshot even after an edit, with Idempotency-Replayed=true; replay does not recheck current date bounds/category ownership. Changed payload409 IDEMPOTENCY_CONFLICT. Deleted original leaves a409 IDEMPOTENCY_REPLAY_UNAVAILABLE tombstone. Without a key, a repeated request can create duplicates. Keys are account-scoped and retained for the account lifetime.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "New expense or original keyed snapshot. Header absent for a new save, present=true on replay.",
            "headers": {
              "X-Request-Id": { "$ref": "#/components/headers/RequestId" },
              "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transaction" } } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/KeyedConflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/transactions/{id}": {
      "patch": {
        "operationId": "patchTransaction",
        "tags": ["Transactions"],
        "summary": "Edit only supplied expense fields",
        "description": "Strict nonempty JSON object: amount_cents,category_id,spent_on,note only. Unknown or server-owned id/user_id/schedule_id/created_at fields refused400. Omitted fields unchanged. Explicit spent_on=null refused400 (not today's default); valid edited date cannot exceed tomorrow. category_id=null/empty string clears, other values Number-converted and ownership checked. note=null clears; other JSON values String-converted. Valid missing/foreign expense or categoryID404.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionPatchRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/TransactionAccepted" },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "operationId": "deleteTransaction",
        "tags": ["Transactions"],
        "summary": "Delete an own expense",
        "description": "Ownership-scoped delete. Valid missing/foreign ID404; invalid conversion400. Success204 is empty, not JSON. A keyed-save record is retained as a tombstone, so a later replay cannot silently recreate the expense. Request body is ignored by the handler; global JSON parser errors still apply.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/Id" }],
        "responses": {
          "204": {
            "description": "Deleted; no response body and no JSON Content-Type.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }
          },
          "400": { "$ref": "#/components/responses/InvalidRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT returned by registration/login. The application's Authorization header uses case-sensitive Bearer followed by a token. Missing, invalid or expired token returns 401. The three public operations explicitly override this root requirement with security=[]."
      }
    },
    "schemas": {
...
      "ErrorMessage": { "type": "string", "minLength": 1 },
      "Error400": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["VALIDATION_FAILED"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error401": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["UNAUTHORIZED"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error404": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["NOT_FOUND"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "KeyedError409": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["IDEMPOTENCY_CONFLICT", "IDEMPOTENCY_REPLAY_UNAVAILABLE"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
      "Error413": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["PAYLOAD_TOO_LARGE"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "Error500": {
        "type": "object", "additionalProperties": false, "required": ["error"],
        "properties": { "error": { "type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": { "code": { "type": "string", "enum": ["INTERNAL"] }, "message": { "$ref": "#/components/schemas/ErrorMessage" } } } }
      },
...
      "LocalDate": {
        "type": "string",
        "format": "date",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "description": "Real local calendar date YYYY-MM-DD, no timezone conversion. Source validates month/day including leap years and imposes no additional year ceiling within four digits."
      },
      "AmountCents": {
        "type": "integer",
        "minimum": 1,
        "maximum": 100000000,
        "description": "Positive integer EUR cents, inclusive ceiling100000000. JSON numbers only; no numeric-string coercion."
      },
      "CategoryIdInput": {
        "oneOf": [
          { "type": "integer", "nullable": true, "minimum": 1 },
          { "type": "string" },
          { "type": "boolean", "enum": [true] },
          { "type": "array", "items": {} }
        ],
        "description": "Omitted/null/empty string -> no category. Otherwise Number(value) must become a positive integer owned categoryID: foreign/missing404, invalidconversion400. Numeric strings, true (1), and numeric arrays such as [1] or nested single-element arrays can coerce. Array/string alternatives are broad because plain schema cannot express JavaScript Number conversion; this prose rule still applies. Objects, false, zero and invalid numeric conversions are not valid owned IDs."
      },
      "Transaction": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "amount_cents", "category_id", "spent_on", "note", "schedule_id", "created_at"],
        "properties": {
          "id": { "type": "integer", "minimum": 1 },
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "type": "integer", "nullable": true, "minimum": 1 },
          "spent_on": { "$ref": "#/components/schemas/LocalDate" },
          "note": { "type": "string", "nullable": true },
          "schedule_id": { "type": "integer", "nullable": true, "minimum": 1 },
          "created_at": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
            "description": "SQLite datetime('now') UTC timestamp with a space, not RFC3339/date-time. Do not parse stored local spent_on through timezone conversion."
          }
        }
      },
      "TransactionCreateRequest": {
        "type": "object",
        "additionalProperties": true,
        "required": ["amount_cents"],
        "properties": {
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "$ref": "#/components/schemas/CategoryIdInput" },
          "spent_on": {
            "type": "string", "nullable": true, "format": "date", "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "description": "Unkeyed omission/null defaults to today. With Idempotency-Key, explicit non-null valid date required. New saves no later than server tomorrow; replay preserves original date. Header-dependent requirement and relative bound are prose rules, not supplied defaults."
          },
          "note": { "$ref": "#/components/schemas/StringifiedInput" }
        },
        "example": { "amount_cents": 1250, "category_id": "1", "spent_on": "2026-09-28", "note": "Synthetic lunch" }
      },
      "TransactionPatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "amount_cents": { "$ref": "#/components/schemas/AmountCents" },
          "category_id": { "$ref": "#/components/schemas/CategoryIdInput" },
          "spent_on": { "$ref": "#/components/schemas/LocalDate" },
          "note": { "$ref": "#/components/schemas/StringifiedInput" }
        },
        "description": "Strict editable fields only; omission preserves, explicit null date invalid. Note null clears; other non-null JSON uses String(value). Edited date cannot exceed tomorrow."
      },
      "TransactionList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "total"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Transaction" } },
          "total": { "type": "integer", "minimum": 0, "description": "Matching row count before limit/offset, not this page's length." }
        }
      },
...
      "StringifiedInput": {
        "description": "For category icon or transaction note: omitted/null -> null on creation. Any non-null JSON value becomes String(value), including booleans, numbers, arrays and objects; no length ceiling is enforced. PATCH omission leaves the field unchanged; explicitnull clears. Schemas enumerate JSON types rather than inventing string-only input validation.",
        "oneOf": [
          { "type": "string", "nullable": true },
          { "type": "number" },
          { "type": "boolean" },
          { "type": "array", "items": {} },
          { "type": "object", "additionalProperties": true }
        ]
      },
...
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Application uses Number(raw) and requires a positive integer, not a digits-only string. Numeric representations such as 1e0 can coerce. Invalid conversion ->400; valid but missing/foreign owned category, transaction or schedule ID ->404."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "pattern": "^[A-Za-z0-9-]{8,64}$" },
        "description": "Only expense POST supports it. Account-scoped lifetime operation key. Explicit non-null spent_on required. Same canonical payload replays201, changed payload409, deleted original409. No key means no deduplication guarantee."
      },
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "A lowercase UUID matching the response header pattern is echoed. Any other value, including uppercase or non-UUID text, is replaced by a generated UUID; it does not cause 400."
      }
    },
    "headers": {
      "IdempotencyReplayed": {
        "description": "Present with value true only on a successful keyed expense replay; absent on a new save. Conditional presence must be tested by the runner, not treated as globally required.",
        "schema": { "type": "string", "enum": ["true"] }
      },
      "RequestId": {
        "description": "Always present. Caller lowercase UUID is echoed; otherwise a new UUID is generated. The x-required extension is for the later response validator, not an OpenAPI required keyword.",
        "x-required": true,
        "schema": {
          "type": "string",
          "format": "uuid",
          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
        }
      },
...
    },
    "responses": {
...
      "TransactionList": {
        "description": "Own filtered expense page and total matching count.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionList" } } }
      },
      "TransactionAccepted": {
        "description": "Edited own expense with persisted fields.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Transaction" } } }
      },
      "KeyedConflict": {
        "description": "IDEMPOTENCY_CONFLICT for changed canonical payload or IDEMPOTENCY_REPLAY_UNAVAILABLE when the original expense was deleted.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KeyedError409" } } }
      },
...
      "NotFound": {
        "description": "NOT_FOUND: no requested category, transaction or schedule belongs to the account. Missing and foreign owned IDs both return404, not403.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error404" } } }
      },
...
      "InvalidRequest": {
        "description": "VALIDATION_FAILED: invalid values, date range, ID conversion or PATCH shape, or malformed JSON rejected before a handler. Login bad/missing credential fields instead use401. The global parser also applies before auth and GET handlers.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error400" } } }
      },
      "Unauthorized": {
        "description": "UNAUTHORIZED: login uses the same E-mail or password is incorrect message for absent/non-string/rejected credentials. Protected mounts refuse absent/malformed bearer headers or invalid/expired tokens with their authentication message. No account-existence disclosure.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error401" } } }
      },
...
      "PayloadTooLarge": {
        "description": "PAYLOAD_TOO_LARGE: the global Express JSON parser's default 100 KiB body limit is exceeded.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error413" } } }
      },
...
      "InternalError": {
        "description": "INTERNAL: Unexpected server error. Documented possible unhandled failure; not induced or observed by this document-validation slice.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error500" } } }
      }
    }
  }
}